requirements.txt.txt
Flask==2.3.3
flask-cors==4.0.0
pyTelegramBotAPI==4.12.0
Pillow==10.0.0
requests==2.31.0
python-dotenv==1.0.0
flask-sqlalchemy==3.0.5
flask-mail==0.9.1
itsdangerous==2.1.2
authlib==1.2.1